REF: Moves the validators to their own folder
