+++ src/controllers/tests/index.js
@@ -4,9 +4,3 @@ require('./jwt.test')
 require('./password.test')
 require('./pushpin.test')
 require('./utils.test')
-require('./validateActivateUser.test')
-require('./validateAuth.test')
-require('./validateLogin.test')
-require('./validateReenableUser.test')
-require('./validateUser.test')
-require('./validationErrorSchema.test')
