+++ src/validators/tests/index.js
@@ -1,0 +1,8 @@
+'use strict'
+
+require('./validateActivateUser.test')
+require('./validateAuth.test')
+require('./validateLogin.test')
+require('./validateReenableUser.test')
+require('./validateUser.test')
+require('./validationErrorSchema.test')
